requirements.txt
Flask
Flask-SQLAlchemy
Flask-Marshmallow
psycopg2-binary
PyJWT
pytest
boto3
marshmallow-sqlalchemy
Flask-Migrate
flask-swagger-ui
python-dotenv